Job queue, and some tutorial stuff
I first went through the tutorial on the Beanstalk deployment I
provided the link to on Slack. I changed code where I could, eventually
stopping because I wasn’t setting up an AWS account. This was mainly to
become more familiar with the tutorial.

Second, I looked at the queue documentation for Laravel. I wanted to
setup a job queue to handle the uploading and Ascii conversion, so for
deployment, there would be a dedicated web server, and the rest of the
servers would handle the job queue tasks (Which is setup in the
tutorial, as long as you have the job queue). I coded a little to
handle the uploading for the user’s computer, so check if I’m doing it
right, then add the changes for the YouTube upload, and the ascii js
script (if possible). (See video controller, and app/Jobs)